% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{get_ate}
\alias{get_ate}
\title{Calculate Average Treatment Effects}
\usage{
get_ate(y, w, reference = 0)
}
\arguments{
\item{y}{A vector of outcomes (i.e., dependent variable)}

\item{w}{An integer vector of treatment assignments}

\item{reference}{The value of w that should be taken to represent the control condition.  Defaults to 0.}
}
\description{
This function calculates the average treatment effects against a 
reference (i.e., control) condition from a vector of treatment 
assignments and an outcome.
}
\examples{
## simulate treatment assignments and outcome
n <- 100
w <- ifelse(runif(n) > .5, 1, 0)
y <- ifelse(runif(n) + .25 * w > .5, 1, 0)

get_ate(y, w)


}
