% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{getImps_}
\alias{getImps_}
\title{Variable importances from orf objects.}
\usage{
getImps_(orfModel)
}
\arguments{
\item{orfModel}{List representing an orf model.}
}
\description{
Workhorse function used to get variable importances from online random forest objects.  Use the function
\code{\link{get_importance}} instead.
}
\seealso{
\code{\link{get_importance}} for making predictions from the orf object.
}
