% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{is.orf}
\alias{is.orf}
\title{Check if an object is an Online Random Forest}
\usage{
is.orf(x)
}
\arguments{
\item{x}{Object to test}
}
\description{
This function checks the class of an object to see if it is an online random forest
}
\examples{
## initialize the model object
orfmod <- init_orf(numClasses = 2, numFeatures = 10, numRandomTests = 2,
                   counterThreshold = 10, maxDepth = 5, numTrees = 10,
                   numEpochs = 1)
                   
is.orf(orfmod)
}
