% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{orf}
\alias{orf}
\title{Online Random Forest}
\usage{
orf(x, y, orfModel, trainModel = TRUE)
}
\arguments{
\item{x}{A matrix of features on which to generate predictions.}

\item{y}{Vector of classes.  Multiclass classification is supported.  Must be integers.}

\item{orfModel}{List representing an orf model.}

\item{trainModel}{Should the forest be trained on the new data?
Defaults to TRUE.  Useful for testing.}
}
\description{
Workhorse function used to fit an online random forest object.  Use the functions
\code{\link{init_orf}} and \code{\link{train_orf}} or \code{\link{online_random_forest}} 
instead.
}
\seealso{
\code{\link{init_orf}} for initializing the orf object, \code{\link{train_orf}} for training the orf object
}
