% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{predict_orf}
\alias{predict_orf}
\title{Predict Online Random Forest}
\usage{
predict_orf(model, x, iteAll = FALSE)
}
\arguments{
\item{model}{An initialized online random forest object created by the `init_orf()` function and ideally trained with data using the `train_orf()` function}

\item{x}{A matrix of features on which to generate predictions.  Must be equal to `numFeatures` when forest was initialized.}

\item{iteAll}{If a causal forest, should all ITE estimates be returned?  Default=FALSE.}
}
\description{
This function trains an initialized online random forest (or causal online random forest).
}
\examples{
## simulate a data point with 10 columns
x <- matrix(runif(10), nrow=1)

## initialize the model object
orfmod <- init_orf(numClasses = 2, numFeatures = 10, numRandomTests = 2, counterThreshold = 10, maxDepth = 5, numTrees = 10, numEpochs = 1)

## train the model with the data
orfmod <- train_orf(model = orfmod, x = x, y=0)

## make predictions on new data
x2 <- matrix(runif(10), nrow=1)
p <- predict_orf(orfmod, x)

}
\keyword{causal}
\keyword{forest}
\keyword{forest,}
\keyword{incremental}
\keyword{learning,}
\keyword{online}
\keyword{out-of-core}
\keyword{random}
