% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{predictOrf}
\alias{predictOrf}
\title{Predictions from Online Random Forests}
\usage{
predictOrf(x, orfModel, allTrees = FALSE)
}
\arguments{
\item{x}{A matrix of features on which to generate predictions.}

\item{orfModel}{List representing an orf model.}

\item{allTrees}{Should all prediction estimates be returned? If TRUE, will 
return a matrix `yHatAllTrees` with one column per tree. If a causal forest, 
will also return a matrix `tauHatAllTrees`. Default=FALSE.}
}
\description{
Workhorse function used to make predictions from online random forest objects.  Use the function
\code{\link{predict.orf}} instead.
}
\seealso{
\code{\link{predict.orf}} for making predictions from the orf object.
}
