% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{causal_online_random_forest}
\alias{causal_online_random_forest}
\title{Causal Online Random Forest}
\usage{
causal_online_random_forest(x, y, treat, numRandomTests, counterThreshold,
  maxDepth, numTrees, numEpochs, type = "classification",
  method = "gini", causal = TRUE, findTrainError = FALSE,
  verbose = FALSE, trainModel = TRUE)
}
\arguments{
\item{x}{A matrix of features on which to generate predictions.  Must be equal to `numFeatures` when forest was initialized.}

\item{y}{Vector of classes.  Multiclass classification is supported.  Must be integers.}

\item{treat}{Vector of treatment assignment.  Must be 0 or 1.}

\item{numRandomTests}{Number of random tests, i.e., number of features to be selected at each node for evaluation.}

\item{counterThreshold}{Threshold for number of observations before a split will occur.}

\item{maxDepth}{Maximum depth for each tree.}

\item{numTrees}{Number of trees in the forest.}

\item{numEpochs}{Number of epochs for processing during each training step.}

\item{type}{Type of Random Forest.  Only `classification` is implemented at this time.}

\item{method}{Method used for determining if a node should split.  Implemented methods are `gini` for Gini Impurity or `entropy` for entropy.
causal_online_random_forest()}
}
\description{
Create a Causal Online Random Forest and fit on provided data.
}
