% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{get_importance}
\alias{get_importance}
\title{Get Feature Importances}
\usage{
get_importance(model)
}
\arguments{
\item{model}{An initialized online random forest object created by the `init_orf()` function and ideally trained with data using the `train_orf()` function}
}
\description{
This function gets the feature importances from an Online Random Forest
}
\keyword{causal}
\keyword{forest}
\keyword{forest,}
\keyword{incremental}
\keyword{learning,}
\keyword{online}
\keyword{out-of-core}
\keyword{random}
