% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orf_functions.R
\name{orf_cv}
\alias{orf_cv}
\title{Online Random Forest Cross Validation}
\usage{
orf_cv(x, y, numClasses, numRandomTests, counterThreshold, maxDepth,
  numTrees, numEpochs, nfolds, type = "classification",
  method = "gini")
}
\arguments{
\item{x}{A matrix of features on which to generate predictions.  Must be equal to `numFeatures` when forest was initialized.}

\item{y}{Matrix of classes with one column.  Multiclass classification is supported.  Must be integers and a matrix.}

\item{numClasses}{Number of classes expected to be classified.}

\item{numRandomTests}{Number of random tests, i.e., number of features to be selected at each node for evaluation.}

\item{counterThreshold}{Threshold for number of observations before a split will occur.}

\item{maxDepth}{Maximum depth for each tree.}

\item{numTrees}{Number of trees in the forest.}

\item{numEpochs}{Number of epochs for processing during each training step.}

\item{nfolds}{Number of folds for cross validation.}

\item{type}{Type of Random Forest.  Only `classification` is implemented at this time.}

\item{method}{Method used for determining if a node should split.  Implemented methods are `gini` for Gini Impurity or `entropy` for entropy.}
}
\description{
This function runs cross validation with a Causal ORF over a dataset to
return out of sample predictions and accuracy.
hyperparameters.
}
\details{
Warning: the cross validation procedure permutes the order of the observations in the output.
}
\examples{
\dontrun{
x <- rnorm()
}
}
